Update compile
Hey, I'm having a bad time deploying my webpack project to heroku... I tries to find a "webpack" folder, and it doesn't work. It also says that the command "npm" is not found. I think you meant to tell it to search the "webpack.config.js" file. This is my proposal, please return to me ASAP!